+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./App.scss";
-import ReturnInterval from "./components/ReturnInterval";
+import ReturnInterval from "./components/ReturnDateSelector";
 
 function App() {
   return (
